refactor: move try_send_msg over to utils
